fix(preprocess): fix dtype in sf.read() to save memory and fix preprocess_resample (#132)

+++ src/so_vits_svc_fork/preprocess_split.py
@@ -20,7 +20,7 @@ def _process_one(
     hop_seconds: float = 0.1,
 ):
     try:
-        audio, sr = sf.read(input_path)
+        audio, sr = sf.read(input_path, dtype="float32")
     except Exception as e:
         LOG.warning(f"Failed to read {input_path}: {e}")
         return
